Put common test code in __init__.py
